Add fmtb to print binary numbers as string
The format string "%.2b" is only available when using the glibc and it
is not portable. Replace this by a new function: fmtb; that fills a
given string buffer with the binary reprenstation of a given number.
The fmtb function returns a pointer into the string buffer that can be
directly passed to printf to be printed.