Rework the workaround for device CH569
Some device running with BTVER 2.7 do send the isp key checksum as
previous versions. The only device tested with which do not send
the checksum but zero is CH569, which has a BTVER of 2.7.

The BTVER version number is likely not relevent on whether the
checksum is sent or zero is sent.

Thus rework the workaround to expect zero, instead of the isp key
checksum, on all CH56x devices.